--- a/pkg/remote/codec/thrift/thrift_test.go
+++ b/pkg/remote/codec/thrift/thrift_test.go
@@ -32,7 +32,7 @@ import (
 )
 
 var (
-	payloadCodec = &thriftCodec{}
+	payloadCodec = &thriftCodec{CodecConfig: CodecConfig{true, true, true, true}}
 	svcInfo      = mocks.ServiceInfo()
 )
 
@@ -155,6 +155,90 @@ func BenchmarkNormalParallel(b *testing.B) {
 			}
 		}
 	})
+}
+
+type MockFrugalTagReq struct {
+	Msg     string            `frugal:"1,default,string"`
+	StrMap  map[string]string `frugal:"2,default,map<string:string>"`
+	StrList []string          `frugal:"3,default,list<string>"`
+}
+
+type MockNoTagArgs struct {
+	Req *MockFrugalTagReq
+}
+
+func initNoTagSendMsg(tp transport.Protocol) remote.Message {
+	var _args MockNoTagArgs
+	ink := rpcinfo.NewInvocation("", "mock")
+	ri := rpcinfo.NewRPCInfo(nil, nil, ink, nil, nil)
+	msg := remote.NewMessage(&_args, svcInfo, ri, remote.Call, remote.Client)
+	msg.SetProtocolInfo(remote.NewProtocolInfo(tp, svcInfo.PayloadCodec))
+	return msg
+}
+
+type MockFrugalTagArgs struct {
+	Req *MockFrugalTagReq `frugal:"1,default,MockFrugalTagReq"`
+}
+
+func initFrugalTagSendMsg(tp transport.Protocol) remote.Message {
+	var _args MockFrugalTagArgs
+	_args.Req = &MockFrugalTagReq{
+		Msg:     "MockReq",
+		StrMap:  map[string]string{"0": "0", "1": "1", "2": "2"},
+		StrList: []string{"0", "1", "2"},
+	}
+	ink := rpcinfo.NewInvocation("", "mock")
+	ri := rpcinfo.NewRPCInfo(nil, nil, ink, nil, nil)
+	msg := remote.NewMessage(&_args, svcInfo, ri, remote.Call, remote.Client)
+	msg.SetProtocolInfo(remote.NewProtocolInfo(tp, svcInfo.PayloadCodec))
+	return msg
+}
+
+func initFrugalTagRecvMsg() remote.Message {
+	var _args MockFrugalTagArgs
+	ink := rpcinfo.NewInvocation("", "mock")
+	ri := rpcinfo.NewRPCInfo(nil, nil, ink, nil, rpcinfo.NewRPCStats())
+	msg := remote.NewMessage(&_args, svcInfo, ri, remote.Call, remote.Server)
+	return msg
+}
+
+func TestFrugal(t *testing.T) {
+	ctx := context.Background()
+
+	// encode client side
+	// MockNoTagArgs cannot be marshaled
+	sendMsg := initNoTagSendMsg(transport.TTHeader)
+	out := remote.NewWriterBuffer(256)
+	err := payloadCodec.Marshal(ctx, sendMsg, out)
+	test.Assert(t, err != nil)
+
+	// MockFrugalTagArgs can be marshaled by frugal
+	sendMsg = initFrugalTagSendMsg(transport.TTHeader)
+	out = remote.NewWriterBuffer(256)
+	err = payloadCodec.Marshal(ctx, sendMsg, out)
+	test.Assert(t, err == nil, err)
+
+	// decode server side
+	recvMsg := initFrugalTagRecvMsg()
+	buf, err := out.Bytes()
+	recvMsg.SetPayloadLen(len(buf))
+	test.Assert(t, err == nil, err)
+	in := remote.NewReaderBuffer(buf)
+	err = payloadCodec.Unmarshal(ctx, recvMsg, in)
+	test.Assert(t, err == nil, err)
+
+	// compare Args
+	sendReq := (sendMsg.Data()).(*MockFrugalTagArgs).Req
+	recvReq := (recvMsg.Data()).(*MockFrugalTagArgs).Req
+	test.Assert(t, sendReq.Msg == recvReq.Msg)
+	test.Assert(t, len(sendReq.StrList) == len(recvReq.StrList))
+	test.Assert(t, len(sendReq.StrMap) == len(recvReq.StrMap))
+	for i, item := range sendReq.StrList {
+		test.Assert(t, item == recvReq.StrList[i])
+	}
+	for k := range sendReq.StrMap {
+		test.Assert(t, sendReq.StrMap[k] == recvReq.StrMap[k])
+	}
 }
 
 func TestException(t *testing.T) {