#### What type of PR is this?

feat

#### What this PR does / why we need it (English/Chinese):

en: Use Frugal as default Thrift En/Decoder
zh: 使用 frugal 作为默认的 Thrift 编解码器

#### Which issue(s) this PR fixes:

none